refactor: move lifecycle module to orchestrator
Moves lifecycle phase resolution, execution, content hash tracking,
and the WaitForPhase enum from cella-cli into cella-orchestrator.
Also moves lifecycle_entries_for_phase (metadata/config fallback logic).

The CLI's lifecycle.rs becomes a thin bridge layer that converts
Progress → ProgressSender for each orchestrator call.
