+++ crates/cella-orchestrator/Cargo.toml
@@ -7,6 +7,7 @@ license.workspace = true
 [dependencies]
 cella-docker.workspace = true
 cella-features.workspace = true
+cella-git.workspace = true
 hex.workspace = true
 serde_json.workspace = true
 sha2.workspace = true
