+++ crates/cella-orchestrator/src/lib.rs
@@ -7,6 +7,7 @@
 pub mod config;
 pub mod error;
 pub mod image;
+pub mod lifecycle;
 pub mod progress;
 pub mod result;
